Fix F811 rule, move fixtures to conftest.py
Moved fixtures from test_utils module to a special conftest.py file that
is grabbed automatically by pytest. This fixes the F811 Flake8 rule
complaining about test function argument names clashing with the pytest
fixture function name.
