+++ test_host_dedup_logic.py
@@ -4,7 +4,7 @@ from api.host import find_existing_host
 from app import db
 from app.models import Host
 from pytest import mark
-from test_utils import flask_app_fixture  # noqa: 401
+
 
 ACCOUNT_NUMBER = "000102"
