+++ test_utils.py
@@ -1,9 +1,7 @@
 import contextlib
 import os
-import pytest
 import unittest.mock
 
-from app import create_app, db
 from app.models import Host
 
 
@@ -33,21 +31,3 @@ def rename_host_table_and_indexes():
     for index in Host.__table_args__:
         if temp_table_name_suffix not in index.name:
             index.name = index.name + temp_table_name_suffix
-
-
-@pytest.fixture
-def flask_app_fixture():
-    rename_host_table_and_indexes()
-
-    app = create_app(config_name="testing")
-
-    # binds the app to the current context
-    with app.app_context() as ctx:
-        # create all tables
-        db.create_all()
-        ctx.push()
-        yield app
-        ctx.pop
-
-        db.session.remove()
-        db.drop_all()
